stress-yield: cast max_ops_per_yielder to double for double comparison

+++ stress-yield.c
@@ -301,7 +301,7 @@ static int stress_yield(stress_args_t *args)
 					pr_fail("%s: sched_yield failed, errno=%d (%s)\n",
 						args->name, errno, strerror(errno));
 				}
-			} while (stress_continue_flag() && (!max_ops_per_yielder || (metrics[i].count < max_ops_per_yielder)));
+			} while (stress_continue_flag() && (!max_ops_per_yielder || (metrics[i].count < (double)max_ops_per_yielder)));
 			_exit(EXIT_SUCCESS);
 		}
 	}
